@@ -4,12 +4,12 @@ import com.teste.demo.exception.TransactionException;
 import com.teste.demo.model.Error;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.handler.HandlerExceptionResolverComposite;
 
 import javax.servlet.http.HttpServletRequest;
+
 
 @ControllerAdvice
 public class WebHandlerException extends HandlerExceptionResolverComposite {
